tmp: 사용자 필수 확인 체크리스트 작업 중 임시 커밋
