+++ src/main/java/com/example/dnd_13th_9_be/checklist/application/port/UserRequiredItemQueryPort.java
@@ -1,0 +1,6 @@
+package com.example.dnd_13th_9_be.checklist.application.port;
+
+public interface UserRequiredItemQueryPort {
+    boolean existsByUserIdAndItemId(Long userId, Long  itemId);
+
+}
