+++ src/main/java/com/example/dnd_13th_9_be/checklist/persistence/repository/UserRequiredItemRepository.java
@@ -1,0 +1,9 @@
+package com.example.dnd_13th_9_be.checklist.persistence.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRequiredItemRepository extends JpaRepository {
+
+}
